build: add bundle analyzer & source maps, implement code splitting for forms & dialogs
