+++ package.json
@@ -59,6 +59,8 @@
     "typescript": "~5.7.2",
     "typescript-eslint": "^8.26.1",
     "vite": "^6.3.1",
+    "vite-bundle-analyzer": "^1.0.0",
     "vitest": "^3.2.4"
-  }
+  },
+  "sideEffects": false
 }
