+++ src/components/tracks/item/TrackActions.tsx
@@ -5,12 +5,13 @@ import {
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
 import { MoreVertical, Pause, Play } from 'lucide-react';
-import { useState } from 'react';
-import TrackFileDialog from './TrackFileDialog';
+import { lazy, Suspense, useState } from 'react';
 import { type TTrack } from '@/lib/schemas';
 import { Button } from '@/components/ui/button';
 import { Checkbox } from '@/components/ui/checkbox';
 import { usePlayerStore } from '@/store/usePlayerStore';
+
+const TrackFileDialog = lazy(() => import('./TrackFileDialog'));
 
 interface ITrackActionsProps {
   isChecked: boolean;
@@ -83,13 +84,15 @@ function TrackActions({
         onClick={e => e.stopPropagation()}
       />
 
-      <TrackFileDialog
-        isOpen={isModalOpen}
-        onClose={closeModal}
-        track={track}
-        onFileUpload={onFileUpload}
-        onDeleteFile={onDeleteFile}
-      />
+      <Suspense fallback={null}>
+        <TrackFileDialog
+          isOpen={isModalOpen}
+          onClose={closeModal}
+          track={track}
+          onFileUpload={onFileUpload}
+          onDeleteFile={onDeleteFile}
+        />
+      </Suspense>
     </div>
   );
 }
